Extract risk concentration contracts
